refactor(broker): move ID token helpers from internal/server into package broker
ContextWithIDToken, GetIDTokenFromContext and GetIDToken are broker-domain
helpers: they encode the OIDC ID token that the broker's HTTP middleware
puts on the request context and that the aggregator reads back when it
delegates to broker ports. Hosting them under internal/server made
internal/server a vestigial package (3 funcs + a doc) and forced both the
broker HTTP adapter and the aggregator to import internal/server for what
is plainly broker functionality.

Moves token_provider.go -> internal/broker/idtoken.go (package broker)
and removes internal/server entirely. Callers in internal/broker/http and
internal/aggregator switch from server.X to broker.X.
