+++ internal/broker/idtoken_test.go
@@ -1,4 +1,4 @@
-package server
+package broker
 
 import (
 	"context"
